elastic-logs/requirements.txt
elasticsearch==8.12.0
kafka-python==2.0.2
urllib3>=2.0.0